dashboard: `<Sidebar />` color scheme toggle style
